Antoine

See merge request ppii2k22/project2-E8!72

+++ Solveur/src/pattern.h
@@ -23,6 +23,6 @@ void pattern_destroy(pattern* one_pattern);
 
 void patterns_destroy(pattern **patterns);
 
-long base3(int n);
+int base3(int n);
 
 #endif /* PATTERN_H */
